+++ src/lib/util.js
@@ -18,3 +18,5 @@ export function toRes(res, status=200) {
 		res.status(status).json(thing);
 	};
 }
+
+export const isInt = value => !isNaN(parseInt(value));
